part

+++ Channel.cpp
@@ -22,7 +22,7 @@ void	Channel::set_topic(const std::string &topic)
 	this->topic = topic;
 }
 
-void	Channel::set_max_users(const u_int16_t &max_users)
+void	Channel::set_max_users(const size_t &max_users)
 {
 	this->max_users = max_users;
 }
